+++ example/lib/base_api_client_example/base_api_client_example.dart
@@ -2,6 +2,7 @@ import 'package:example/base_api_client_example/domain/entity/user_entity.dart';
 import 'package:example/base_api_client_example/domain/repository/user_repository.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
+import 'package:onix_flutter_core/onix_flutter_core.dart';
 
 class BaseApiClientExample extends StatelessWidget {
   final String title;
@@ -64,7 +65,7 @@ class BaseApiClientExample extends StatelessWidget {
 
   Future<List<UserEntity>?> _getUsers() async {
     final result = await GetIt.I.get<UserRepository>().getUsers();
-    if (result.success) {
+    if (result.isOk) {
       return result.data;
     }