+++ example/lib/base_api_client_example/data/model/default_api_error.dart
@@ -1,0 +1,15 @@
+class DefaultApiError {
+  final String? message;
+  final String? code;
+
+  DefaultApiError({
+    required this.message,
+    required this.code,
+  });
+
+  factory DefaultApiError.fromJson(Map<String, dynamic> json) =>
+      DefaultApiError(
+        message: json["message"] as String?,
+        code: json["code"] as String?,
+      );
+}