+++ example/lib/base_api_client_example/data/repository/user_repository_impl.dart
@@ -1,9 +1,8 @@
-import 'package:example/base_api_client_example/data/defalt_api_error.dart';
+import 'package:example/base_api_client_example/core/mapper/dio_server_error_mapper.dart';
 import 'package:example/base_api_client_example/data/mapper/user_mapper.dart';
 import 'package:example/base_api_client_example/data/source/user_source.dart';
 import 'package:example/base_api_client_example/domain/entity/user_entity.dart';
 import 'package:example/base_api_client_example/domain/repository/user_repository.dart';
-import 'package:onix_flutter_core/onix_flutter_core.dart';
 import 'package:onix_flutter_core_models/onix_flutter_core_models.dart';
 
 class UserRepositoryImpl implements UserRepository {
@@ -24,25 +23,16 @@ class UserRepositoryImpl implements UserRepository {
             )
             .toList();
 
-        return Result.success(users);
+        return Result.ok(users);
       } else {
-        final failure = MapCommonServerError.getServerFailureDetails(
+        final error = DioServerErrorMapper().mapToFailure(
           userResponse,
-          onApiFailure: (error, _) {
-            if (error is DefaultApiError) {
-              return ApiFailure(
-                ServerFailure.response,
-                message: error.message ?? '',
-              );
-            }
-            return ApiFailure(ServerFailure.response);
-          },
         );
-        return Result.error(failure: failure);
+        return Result.error(error: error);
       }
     } catch (e) {
       return Result.error(
-        failure: ApiFailure(ServerFailure.unknown),
+        error: ApiFailure(ServerFailure.unknown),
       );
     }
   }