+++ example/lib/base_api_client_example/data/source/user_source_impl.dart
@@ -1,4 +1,3 @@
-import 'package:example/base_api_client_example/data/defalt_api_error.dart';
 import 'package:example/base_api_client_example/data/model/user_model.dart';
 import 'package:example/base_api_client_example/data/source/user_source.dart';
 import 'package:onix_flutter_core/onix_flutter_core.dart';
@@ -7,19 +6,17 @@ class UserSourceImpl implements UserSource {
   static const String _users = '/users';
 
   final ApiClient _apiClient;
-  final DioRequestProcessor _dioRequestProcessor;
+  final RequestProcessor _requestProcessor;
 
-  const UserSourceImpl(this._apiClient, this._dioRequestProcessor);
+  const UserSourceImpl(this._apiClient, this._requestProcessor);
 
   @override
   Future<DataResponse<UserModelList>> getUsers() {
-    return _dioRequestProcessor.processRequest(
-        onRequest: () => _apiClient.client.get(_users),
-        onResponse: (response) {
-          return UserModelList.fromJson(response.data);
-        },
-        onCustomError: (response) {
-          return DefaultApiError(message: response?.data.toString());
-        });
+    return _requestProcessor.processRequest(
+      onRequest: () => _apiClient.client.get(_users),
+      onParse: (response) {
+        return UserModelList.fromJson(response.data);
+      },
+    );
   }
 }