+++ lib/onix_flutter_core.dart
@@ -1,19 +1,34 @@
 library onix_flutter_core;
 
-export 'src/data/local/prefs/base_preferences.dart';
+export 'src/data/local/prefs/shared_preferences_storage.dart';
+export 'src/data/local/base/key_value_storage.dart';
+export 'src/data/local/base/key_value_reloadable_storage.dart';
 export 'src/data/remote/base/base_api_client.dart' show BaseApiClient;
 export 'src/data/remote/base/http_status.dart' show HttpStatus;
-export 'src/data/remote/base/map_common_server_error.dart'
-    show MapCommonServerError;
-export 'src/data/remote/connection_checker/connection_checker.dart';
+export 'src/data/remote/base/server_error_mapper.dart'
+    show ServerErrorMapper;
+export 'src/data/remote/base/connection_checker.dart';
 export 'src/data/remote/dio/api_client.dart' show ApiClient;
 export 'src/data/remote/dio/dio_client_module.dart' show DioClientModule;
-export 'src/data/remote/dio/dio_request_processor/dio_request_processor.dart'
-    show DioRequestProcessor, OnRequest, OnResponse;
-export 'src/data/remote/dio/dio_request_processor/dio_request_processor_impl.dart';
+
+export 'src/data/remote/base/processor/request_processor.dart'
+    show RequestProcessor, OnRequest, OnParse;
+export 'src/data/remote/dio/internal_dio_request_processor.dart'
+    show InternalDioRequestProcessor;
+export 'src/data/remote/base/processor/error_processor.dart'
+    show ErrorProcessor, OnCustomError;
+export 'src/data/remote/dio/internal_dio_error_processor.dart'
+    show InternalDioErrorProcessor;
+export 'src/data/remote/connection_checker/always_have_connection.dart'
+    show AlwaysHaveConnection;
+export 'src/data/remote/connection_checker/mobile_connection_checker.dart'
+    show MobileConnectionChecker;
+export 'src/data/remote/base/retry_policy.dart'
+    show RetryPolicy;
+
+
 export 'src/data/remote/dio/interceptor/cache_interceptor.dart';
 export 'src/data/remote/dio/params/api_client_params.dart' show ApiClientParams;
 export 'src/domain/common/converter/mapper.dart';
-export 'src/domain/entity/common/data_response.dart' show DataResponse;
+export 'src/domain/entity/common/data_response.dart';
 export 'src/domain/entity/common/operation_status.dart' show OperationStatus;
-export 'src/domain/entity/common/result.dart' show Result;